backend/requirements.txt
Flask==3.0.3
Flask-CORS==5.0.0
python-dotenv==1.0.1
PyPDF2==3.0.1
google-generativeai==0.8.3
faiss-cpu==1.8.0
numpy==1.26.4
werkzeug==3.0.6
packaging
